More slot fixes.

+++ Trunk/src/hearthstone-world/Container.cpp
@@ -131,7 +131,7 @@ bool Container::HasItems()
 	return false;
 }
 
-bool Container::AddItem(int8 slot, Item* item)
+bool Container::AddItem(int16 slot, Item* item)
 {
 	if (slot < 0 || (uint32)slot >= GetProto()->ContainerSlots)
 		return false;
@@ -170,7 +170,7 @@ bool Container::AddItem(int8 slot, Item* item)
 	return true;
 }
 
-void Container::SwapItems(int8 SrcSlot, int8 DstSlot)
+void Container::SwapItems(int16 SrcSlot, int16 DstSlot)
 {
 	Item* temp;
 	if( SrcSlot < 0 || SrcSlot >= (int8)m_itemProto->ContainerSlots )
@@ -230,7 +230,7 @@ void Container::SwapItems(int8 SrcSlot, int8 DstSlot)
 	}
 }
 
-Item* Container::SafeRemoveAndRetreiveItemFromSlot(int8 slot, bool destroy)
+Item* Container::SafeRemoveAndRetreiveItemFromSlot(int16 slot, bool destroy)
 {
 	if (slot < 0 || (uint32)slot >= GetProto()->ContainerSlots)
 		return NULLITEM;
@@ -260,7 +260,7 @@ Item* Container::SafeRemoveAndRetreiveItemFromSlot(int8 slot, bool destroy)
 	return pItem;
 }
 
-bool Container::SafeFullRemoveItemFromSlot(int8 slot)
+bool Container::SafeFullRemoveItemFromSlot(int16 slot)
 {
 	if (slot < 0 || (uint32)slot >= GetProto()->ContainerSlots)
 		return false;
@@ -314,7 +314,7 @@ bool Container::AddItemToFreeSlot(Item* pItem, uint32 * r_slot)
 }
 
 
-void Container::SaveBagToDB(int8 slot, bool first, QueryBuffer * buf)
+void Container::SaveBagToDB(int16 slot, bool first, QueryBuffer * buf)
 {
 	SaveToDB(INVENTORY_SLOT_NOT_SET, slot, first, buf);
 
@@ -326,5 +326,3 @@ void Container::SaveBagToDB(int8 slot, bool first, QueryBuffer * buf)
 		}
 	}
 }
-
-
